Make sure line endings are normalized to LF
We need a .gitignore file to make sure that line endings get normalized
to LF. My last two commits got distorted because of this - they were
very small changes that are showing up as big changes because of the
line ending change. Same goes for the draw.r and shape.r files which I
didn't change at all.