Fix build: rename obj/ to objs/
The makefile currently depends on an objs/ directory to exist, but we
ship an obj/ directory. Fix the immediate issue by renaming obj/ to
objs/. Adapted the README accordingly.

Ultimately, (a) obj/ probably is the better name and (b) the build
should create it if it does not already exist.