Temporary workaround for resolve/extend/only crash (http://issue.cc/r3/1865)
RESOLVE/extend/only crashes sometimes, under obsure and (so far)
unpredictable circumstances, but unfortunately not rarely. This crash
only affects modules which export stuff, and not consistently, but often
enough that modules are unusable.

We have to patch this fast. This patch comments out all calls to this
function with these options, replacing them with semantically equivalent
calls. A little more overhead, but at least it doesn't crash.  We can
revert these once someone figures out how to fix the bug in RESOLVE.