Better comments for sys/load-header
The comments of sys/load-header were slightly outdated, and not detailed
enough for people to understand how it works. They also didn't mention
the two embedding methods, nor were they correct in asserting that
checksums didn't work for script-in-a-block.

The new comments mention a gotcha about the 'content header, the
embedded script methods, multi-script support, and the reason that the
body of a script-in-a-block is returned decoded.