Multi-script support, and better support for length-specified embedding
The sys/load-header function has supported both length-specified and
script-in-a-block embedding, but only the latter has been supported by
the functions that call it, such as load.

This adds support for length-specified embedded scripts to load,
load-module, and load-ext-module. It also makes load-module and
load-ext-module also return the end position of their scripts, which
makes multi-script loaders possible. It also adds comments to
load-boot-exts to explain their limitations, and makes sure that the
changes to load-ext-module don't bother it.