Prevent misunderstandings like http://issue.cc/r3/1913
By design, TRANSCODE/error inserts the error, that it would otherwise
trigger, in the place where it would insert a value if there was no
error. This provides valuable context for the error, and can be used to
help in recovery of the transcoding process if you want to. This doc
string tweak makes that more clear so there will be less confusion about
this.