Add support for APPLY :DO [:function]
APPLY :DO [:function] handles the first argument block (the one to DO)
but doesn't handle the second (the one to the function). This change
will allow you to handle the second argument block with a chained
call to APPLY, or to not do the reval at all if you like.

See http://issue.cc/r3/2135 for details.