Make APPLY secure against RETURN/redo (#1949)
You can break through APPLY security with RETURN/redo or native code
equivalent to that, code which is not buggy or bad in itself.

The APPLY native should clear the OPTS_REVAL flag from its return value
before it returns. The native itself, not Apply_Block in case it is used
elsewhere internally in cases where OPTS_REVAL would be OK. It is only
in calls to the native APPLY that it is a problem.

Current behavior
>> apply does [return/redo :add] [] 1 1
== 2

Preferred behavior
>> same? :add apply does [return/redo :add] []
== true

See http://issue.cc/r3/1949