Fix PARSE `THRU integer` operation
PARSE's `TO integer` operation positions "at" the given index. In line
with how PARSE's TO/THRU relate otherwise, `THRU integer` should
therefore move to the position "next" to the given index:

    >> parse "abcd" [to 2 (comment {Positions to "bcd".}) "bcd"]
    == true

    >> parse "abcd" [thru 2 (comment {Should position to "cd".}) "cd"]
    == false  ;; Expected: true

This commit adds the missing handling for distinguishing THRU from TO
in the case handling `TO/THRU integer` operations.

This fixes CureCode issue #1965.