Make CLOSURE a FUNCT for closures (#2002)
This would make CLOSURE like FUNCTION, gathering locals, to make
that locals gathering seem like the primary usage model for user
functions in Rebol, to make the language easier to use.

The old CLOSURE function is renamed to CLOS, to be like FUNC for
closures the way the function is already in practice.

See http://issue.cc/r3/2002 and http://issue.cc/r3/1973 for details.