Fix "Cannot find REBOL= definition" for make-make.r on Win32 (0.3.1)
This fixes a minor issue introduced by a99d2d6: when running
`make-make.r` with a target OS ID of `0.3.1` (Win32), you'd get a
"Cannot find REBOL= definition" message.

This was due to the simple makefile definition replacement logic looking
for "REBOL" would stop on the newly introduced "REBOL_TOOL" definition,
print above error message and fail to proceed to and update the real
"REBOL=" definition.

Further, the generated makefile would check for and try to run `r3-make`
on Win32, instead of the arguably more customary `r3-make.exe`.

The bundled `0.4.4` makefile was re-generated to incorporate the change
introduced to make `r3-make.exe` work. For 0.4.4 itself, that change is
purely cosmetic.