FIX: better handling ANSI escape sequences (on Windows)
In longer outputs (> 1024 bytes) it was possible that the output from the STDIO device handler was divided in the middle of an escape sequence, and it was recognized as an invalid. In this commit I made the parser's state persistent so parsing can continue. It can be emulated from console using such a code:

```
>> prin "^[" print "[1mHello"
Hello
>> prin "^[[" print "1mHello"
Hello
>> prin "^[[1" print "mHello"
Hello

```
Where in all cases the printed "Hello" text is bright.