ATRONIX: Fix bias overflow in Remove_Series
bias is 16-bit, and len is 32-bit, so check overflow before modifing
bias.

This fixes the crash:
a: copy ""
insert/dup a #"a" to integer! #10ffff
take/part a 3
take/part a to integer! #f0010

insert/dup a #"b" 10
a: 1 ;force a to recycle
recycle
print a

(cherry picked from commit cc625bebcb6038b9282876954f929c9d80048d2b)