FEAT: making UNSET! as a conditionally "TRUTHY"
related issues:
https://github.com/rebol/rebol-issues/issues/564
https://github.com/rebol/rebol-issues/issues/850

It reverts changes discussed in above issues as I consider this behavior to be more useful and it is compatible with Rebol2 and Red.

My main reason is to be able use PRINT inside ALL blocks without need to worry if these would affect the evaluation.

Like in this case:
```
	if all [
		not headers
		any [
			all [
				d1: find conn/data crlfbin
				d2: find/tail d1 crlf2bin
				print "server using standard content separator of #{0D0A0D0A}"
			]
			all [
				d1: find conn/data #{0A}
				d2: find/tail d1 #{0A0A}
				print "server using malformed line separator of #{0A0A}"
			]
		]
	][ ... ]

```