FEAT: Pass indexes to ARRAY/initial function
When calling ARRAY/initial with a function used to generate the values,
pass the current index at the position the function is being called for,
with each index in a block of sizes provided to create a multidimensional
array being provided as a separate parameter to the function in the order
specified in the sizes block. Uses APPLY so the function can optionally
not take the parameters. Uses an internal function option for efficiency.

See https://github.com/rebol/rebol-issues/issues/2193 for details.

(cherry picked from commit c08295333315c10284b95cede880c4a15bfe3ae3)