FIX: remove CRLF on Windows and LF on Posix platforms from tail of INPUT line
INPUT function was using TRIM/with newline, but that was not enough on Windows as it was not removing the CR byte. Also it was quite overkill as there should not be any other newlines in the system/ports/input line.

Also with this change, the output on all platforms should be same in this situation:
```
>> read system/ports/input
foo
== #{666F6F}
```

Before this modification it would return `#{666F6F0D0A}` on Windows and `#{666F6F0A}` on Linux.