CHANGE: ATRONIX: inverting alpha channel values and default white color
Most system are using RGBA with A(alpha) = 255 as fully opaque and 0 as fully transparent. In Rebol2 and R3-alpha it was inverted, which causes many image related native code to flip the value on runtime.

In this state (compatible with Atronix branch), when the alpha value is not provided, it is added as 255 (0.0.0 is same like 0.0.0.255). This is a little bit illogical and question is if not to keep original behavior when setting/accessing image colors (and to be compatible with Rebol2, R3-alpha and Red instead).

This commit also changes default image color to be white instead of black as was introduced in Red language.

Contains also fix for: https://github.com/rebol/rebol-issues/issues/2344