FEAT: CHANGE: making SET native to be more compatible with Red-language version
There is new refinement /ONLY which treat source block or object like single value (so the result is like it was in R3-alpha)

The /PAD refinement was replaced with /SOME refinement. The functionality is now reversed: when no refinement is used, it works like if there was /PAD before. If /SOME is used, there is no padding and also target value is not replaced with none value from the source, if there is any.

Related issue: https://github.com/rebol/rebol-issues/issues/2358