FIX/CHANGE: `wide` internal encoding was not used for strings containing chars in range >= 80h and <= FFh
There is no suitable test for it, as there is no function returning internal char size, but with `DUMP` function enabled, it was clearly visible, that before this change:

```
>> dump to-string #{A3}
=> Series #033A8EF0 "-": wide: 1  size: 8      bias: 0 tail: 1 rest: 8 flags: #00000001
03031670: A300                                ..

```
Notice that there is: `wide: 1` in the output above.

Now with the same code it is 2:
```
>> dump to-string #{A3}
=> Series #7C18FAF0 "-": wide: 2  size: 8      bias: 0 tail: 1 rest: 4 flags: #00000002
7E0C4CD8: A3000000                            ....
```

Memory consumption in such a cases may be higher, but it simplifies other manipulation as series which has wide = 1 now cannot hold chars, which needs multibyte encoding.