FIX: errors should not HALT to the console but QUIT to the system
Fixes issue #2215.

It is just a small modification, where in case of script error, the script QUITs instead of HALTing in REPL console. When `system/options/quiet = FALSE`, than there is 3s delay, before automatic quit. During this time user may use CTRL-C to HALT and enter the REPL for in place debugging.

When used `--halt` option, it enters the REPL as it was before without any wait. (Don't know if it is the best way, but I'm not sure if a new option is really needed for forcing REPL in case of errors).