FEAT: updating Windows' prebuild R3 executable
The previous one was over year old, so it's time to update it into up-to-date version. Especially that now it by default does not enter console on error in evaluated script.
(https://github.com/Oldes/Rebol3/commit/61c138334d9fd93ca5760fdf685080f285251193)

Note that it is not fully featured version to keep the executable small.