FEAT: improving HELP output on datatypes.
It also displays different results if used directly on datatype, like:
```
help char!
```

or when used on value, which holds datatype, like in:
```
my-value: char!
help my-value
```

First one displays:
```
CHAR! is a datatype.
It is defined as a 8bit and 16bit character.
It is of the general type scalar.

Found these related words:
  null            char!      #"^@"
  space           char!      #" "
  sp              char!      #" "
  backspace       char!      #"^H"
  bs              char!      #"^H"
  tab             char!      #"^-"
  newline         char!      #"^/"
  newpage         char!      #"^L"
  slash           char!      #"/"
  backslash       char!      #"\"
  escape          char!      #"^["
  cr              char!      #"^M"
  lf              char!      #"^/"
```

while the second one just:
```
MY-VALUE is a datatype of value: char!
```