FIX: code refactoring
Keeping just a `MT_Map` function for making a new `map` value. Years ago I introduced `Copy_Map`, because I was just learning the code structure.. now I'm a little bit more advanced and so know, that `MT_*` are common functions for _making datatype *_ and that the `type` argument of this function is used to specify, how to deal with values, when we want to `copy` the `map`.