FEAT/CHANGE: allowing @name to be a valid email! type
For a compatibility with Red language and also because it may be useful in dialects.
Although `@name` is not a valid real email according RFC, one can construct invalid email values in many ways and so in real use for email apps, there should be a real additional email validation anyway.

Related to: https://github.com/Oldes/Rebol-issues/issues/1962