FEAT/CHANGE: including GIT info in `system/build`
Before this commit, `system/build` was returning just a date, when the executable was compiled.
Now it is an object, with keys `date` and `git`. If the executable is compiled from a valid git repository, `git` contains another object with info about actual `repository`, `branch`, `commit` and it's `message`.

With this info it will be easier to localize possible regressions against older executables.