CHANGE/FIX: rewritten `list-dir` to use recent `dir-tree` function.
The column file list is now removed as it was ugly and not much practical. Now it always display result as a directory tree. Just if there is not specified recursion (/r), the depth is limited to 1. This is backward compatibility breaking change, but easy to fix if needed.

It greatly improve readability of the output (at least for my personal use).
I can imagine, that it could be even more improved (like adding better filtering and or sorting).

I've also added `dir` as an alias to `list-dir` (as it is in Red language).