FEAT: ported Nenad's MySQL driver to be compatible with this Rebol3 version as a module/scheme
It's just a quick rewrite, which replaces original `debug` function in the protocol with system logging and some of required modifications. There is also minnor change in error reporting. It would deserve better rewrite, but I don't have much need for this, so it's left for future or someone else.

It was tested only with this so far:
```
>> mysql: connect-sql mysql://root:rur@localhost/
 [MySQL] Connected: mysql://root@localhost:3306/
>> send-sql mysql "show databases"
== [["information_schema"] ["mysql"] ["performance_schema"] ["sys"] ["test"]]

>> send-sql mysql "create database test"
** mysql error: [ 1007 ] "Can't create database 'test'; database exists"

>> send-sql mysql "drop database test"
== []

>> send-sql/flat mysql "show databases"
== ["information_schema" "mysql" "performance_schema" "sys"]

>> close mysql
 [MySQL] Closed: mysql://root@localhost:3306/
>>
```