FEAT: implemented proper `ask/hide` and `input/hide`
That was not working well on Windows (with enabled virtual terminal processing), now it successfully hides user input.

As a bonus this change contain a way how to turn off read-line console input and enter use the console as a key-stroke (and mouse as well) input. Simple use example is available here:
https://gist.github.com/Oldes/face8a5790df7121a78ba606a3e150f4