FEAT: including first version of higher level `mail` scheme for sending emails
So far it is possible to use Rebol2 like `send` function:
```rebol
import 'mail

; sending just a plain message:
send you@example.com "Hello^/ this is just simple message"

; sending single file:
send you@example.com %path/to/file

; sending more files:
send you@example.com [%path/to/file1 %path/to/file2]

; sending more complete message:
send friends [
    subject: "Invitation"
    message: "Hi friends, check the invitation!"
    attach:  %path/to/invitation
]
```
It is also possible to use the scheme directly, like:
```rebol
write mail:// [
    From:   ["Oldes" oldes@example.com]
    To:     ["Bob" bob@example.com "Eva" eva@example.com]
    subject: "Invitation"
    message: "Hi friends, check the invitation!"
    attach:  %path/to/invitation
]

```
In all cases above is expected, that `system/user` is configured and that there is available `smtp` specification (host/user/password) under `system/users/data`!

This is still just the initial version. So far it works only in sync mode, but the idea is, to be able use it in async mode as well (for example from a server instance).